+++ lib/constants.jsx
@@ -40,7 +40,8 @@ export const CONTRACT_ADDRESSES = {
     GovernanceToken: '0x0cEC1A9154Ff802e7934Fc916Ed7Ca50bDE6844e',
     PrizePoolBuilder: PoolWithMultipleWinnersBuilderMainnet.address,
     GovernanceReserve: '0xdb8E47BEFe4646fCc62BE61EEE5DF350404c124F',
-    MerkleDistributor: '0xBE1a33519F586A4c8AA37525163Df8d67997016f'
+    MerkleDistributor: '0xBE1a33519F586A4c8AA37525163Df8d67997016f',
+    PoolPoolTicket: '0x27d22a7648e955e510a40bdb058333e9190d12d4'
 
     // GovernorAlpha: '0x5e4be8Bc9637f0EAA1A755019e06A68ce081D58F',  // Uniswap Governance for testing
     // GovernanceToken: '0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984' // Uniswap UNI Token  for testing
@@ -73,7 +74,8 @@ export const QUERY_KEYS = {
   usePools: 'usePools',
   prizePoolContractAddresses: 'prizePoolContractAddresses',
   etherscanContractAbi: 'etherscanContractAbi',
-  useProposalVotesTotalPages: 'useProposalVotesTotalPages'
+  useProposalVotesTotalPages: 'useProposalVotesTotalPages',
+  poolPoolBalance: 'poolPoolBalance'
 }
 
 export const ETHERSCAN_API_KEY = process.env.NEXT_JS_ETHERSCAN_API_KEY
