+++ lib/hooks/useTokenHolder.js
@@ -76,6 +76,8 @@ async function getTokenHolder(address, chainId, blockNumber) {
       return EMPTY_TOKEN_HOLDER
     }
 
+    console.log('subgraphData', subgraphData)
+
     const isDelegating = Boolean(subgraphData?.tokenHolder?.delegate?.id)
     const isSelfDelegated =
       address.toLowerCase() === subgraphData?.tokenHolder?.delegate?.id?.toLowerCase()
@@ -83,8 +85,9 @@ async function getTokenHolder(address, chainId, blockNumber) {
     const canVote = Number(subgraphData?.delegate?.delegatedVotes || 0) > 0
     const tokenBalance = subgraphData?.tokenHolder?.tokenBalance || '0'
     const isBeingDelegatedTo =
-      Boolean(subgraphData?.tokenHolder?.delegate?.delegatedVotes) &&
-      subgraphData?.tokenHolder?.delegate?.delegatedVotes > tokenBalance
+      subgraphData?.delegate?.tokenHoldersRepresentedAmount > 2 ||
+      (subgraphData?.delegate?.tokenHoldersRepresentedAmount === 1 &&
+        subgraphData?.delegate?.delegatedVotes !== tokenBalance)
     const tokenHoldersRepresentedAmount = isSelfDelegated
       ? String(Number(subgraphData?.delegate?.tokenHoldersRepresentedAmount) - 1)
       : subgraphData?.delegate?.tokenHoldersRepresentedAmount
